feature(gui): Add a DifferenceGeneratorWrapper class
This class basically wraps lib2/DifferenceGenerator and makes it accessible to the gui.

Sorry for the huge commit:
I also ripped everything out of main.kt
renamed lots of classes
removed code duplications
introduced a new directory structure
